Added files that should be added by Git but were not.
Tidied up code formatting, updated some documentation.

+++ src/main/java/net/boat/industrialhellscape/item/modded_items/JobPaperworkItem.java
@@ -1,22 +1,18 @@
 package net.boat.industrialhellscape.item.modded_items;
 
 
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.particle.MobAppearanceParticle;
-import net.minecraft.client.renderer.entity.ElderGuardianRenderer;
 import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.sounds.SoundEvents;
-import net.minecraft.world.*;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.ai.memory.MemoryModuleType;
-import net.minecraft.world.entity.monster.ElderGuardian;
 import net.minecraft.world.entity.npc.Villager;
 import net.minecraft.world.entity.npc.VillagerData;
 import net.minecraft.world.entity.npc.VillagerProfession;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.entity.schedule.Activity;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.gameevent.GameEvent;
@@ -36,17 +32,12 @@ public class JobPaperworkItem extends Item {
     }
 
     public @NotNull InteractionResult interactLivingEntity(ItemStack stack, Player player, LivingEntity target, InteractionHand hand) {
-        //SaddleItem
-        //AssignProfessionFromJobSite
-        //p_258310_.setVillagerData(p_258310_.getVillagerData().setProfession(p_22464_));
-
         if (target instanceof Villager villager) {
 
             VillagerProfession jobToAcquire = VillagerProfession.NONE;
 
             if (villager.isAlive() && hasThisJobUp.test(villager) && !villager.hasCustomName())  {
 
-                //IF CLIENTSIDE TRUE, VILLAGER REACQUIRES TRADE. BUT IF !CLIENTSIDE TRUE, THIS DOESNT OCCUR. TEST
                 if (!villager.level().isClientSide)  { //IF ON SERVER
 
                     //The order of releasing POIs, setting villager profession, and refreshing their brain matters heavily
@@ -80,10 +71,6 @@ public class JobPaperworkItem extends Item {
 
                 } else { //IF ON CLIENT
                     addParticlesAroundSelf(villager);
-                    //Minecraft.getInstance().gameRenderer.displayItemActivation(stack);
-                    //player.level().addParticle(ParticleTypes.ELDER_GUARDIAN, player.getX(), player.getY(), player.getZ(), 0.0, 0.0, 0.0);
-
-                    //MobAppearanceParticle
                 }
 
             } else if(!villager.level().isClientSide) {
@@ -102,7 +89,7 @@ public class JobPaperworkItem extends Item {
                 double d0 = entity.getRandom().nextGaussian() * 0.02;
                 double d1 = entity.getRandom().nextGaussian() * 0.02;
                 double d2 = entity.getRandom().nextGaussian() * 0.02;
-                entity.level().addParticle(ParticleTypes.HAPPY_VILLAGER, entity.getRandomX((double)1.0F), entity.getRandomY() + (double)1.0F, entity.getRandomZ((double)1.0F), d0, d1, d2);
+                entity.level().addParticle(ParticleTypes.HAPPY_VILLAGER, entity.getRandomX(1.0F), entity.getRandomY() + 1.0, entity.getRandomZ(1.0F), d0, d1, d2);
             }
         }
     }
